feat/refactor(Cart.kt, CartItem.kt, CartDto.kt, CartItemDto.kt): create DTos and models
- create DTOs and models for the cart and cart items
- move them into separate folders
- change import directives accordingly
