+++ src/main/kotlin/ecommerce/dto/CartDto.kt
@@ -1,0 +1,6 @@
+package ecommerce.dto
+
+data class CartDto(
+    val cartId: Long,
+    val userId: Long,
+)
