+++ src/main/kotlin/ecommerce/dto/CartItemDto.kt
@@ -1,0 +1,8 @@
+package ecommerce.dto
+
+class CartItemDto(
+    val cartId: Long,
+    val userId: Long,
+    val productId: Long,
+    val quantity: Int,
+)
