+++ src/main/kotlin/ecommerce/dto/TokenRequest.kt
@@ -1,4 +1,4 @@
-package ecommerce.model
+package ecommerce.dto
 
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.NotBlank
