+++ src/main/kotlin/ecommerce/dto/TokenResponse.kt
@@ -1,3 +1,3 @@
-package ecommerce.model
+package ecommerce.dto
 
 class TokenResponse(val token: String)
