+++ src/main/kotlin/ecommerce/model/CartItem.kt
@@ -1,0 +1,6 @@
+package ecommerce.model
+
+data class CartItem(
+    val product: Product,
+    val quantity: Int,
+)
