+++ src/main/kotlin/ecommerce/repository/CartRepository.kt
@@ -1,0 +1,12 @@
+package ecommerce.repository
+
+import org.springframework.jdbc.core.JdbcTemplate
+import org.springframework.jdbc.core.RowMapper
+import org.springframework.jdbc.core.simple.JdbcClient
+import org.springframework.stereotype.Repository
+
+@Repository
+class CartRepository(private val jdbcClient: JdbcClient) {
+
+
+}
